+++ CHANGELOG.md
@@ -75,6 +75,9 @@
 - tauri: improve performance a little #4812
 - settings: fix: wait for setting to be applied before calling callback #4754
 
+### Removed
+- remove experimental option to disable IMAP IDLE #4991
+
 <a id="1_56_0"></a>
 
 ## [1.56.0] - 2025-03-21